Added spinner animation for loading during form submission #5
