+++ package.json
@@ -42,6 +42,7 @@
     "pg": "^8.8.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-spinners": "^0.13.4",
     "unsplash-js": "^7.0.15"
   },
   "devDependencies": {
