+++ server/public/pages/gym.css
@@ -10,7 +10,7 @@
 @media (min-width: 960px) {
   .gym-main {
     flex-direction: row;
-    align-items: flex-start;
+    align-items: center;
   }
 }
