+++ server/public/scss/pages/gym.scss
@@ -18,7 +18,7 @@
 
   @include lg {
     flex-direction: row;
-    align-items: flex-start;
+    align-items: center;
   }
 }
